% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbR6_add_table.R
\name{add_table}
\alias{add_table}
\alias{dbR6_add_table}
\alias{add_table,dbR6}
\title{add_table}
\usage{
my_object$add_table(from, to, overwrite, append fun, index_row_names)
}
\arguments{
\item{...}{Additional parameters passed to RSQLite::dbWriteTable}

\item{from}{Input table}

\item{to}{Name of the table in database}

\item{overwrite}{Overwrite table if already exists? Default FALSE}

\item{append}{Append rows if the tablealready exists? Default FALSE}

\item{fun}{Function to apply to the table before writing it}

\item{index_row_names}{index row names? Default TRUE}
}
\description{
Add a table to a dbR6  object
}
\examples{
{
data(mtcars2)
my_db <- dbR6$new()
my_db$add_table(mtcars, "mtcars_db")
}
}
