% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aux_format_object_size}
\alias{aux_format_object_size}
\title{generate a formatted number with corresponding units for plyr object}
\usage{
aux_format_object_size(x)
}
\description{
generate a formatted number with corresponding units for plyr object
}
