% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbR6_copy_table_structure.R
\name{copy_table_structure}
\alias{copy_table_structure}
\alias{dbR6_copy_table_structure}
\alias{copy_table_structure,dbR6}
\title{copy_table_structure}
\usage{
dbR6_copy_table_structure(...)
}
\description{
copy_table_structure
}
