% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbR6_copy_table_structure.R
\name{dbR6_copy_table_structure}
\alias{dbR6_copy_table_structure}
\title{dbR6_copy_table_structure__}
\usage{
dbR6_copy_table_structure(...)
}
\description{
dbR6_copy_table_structure__
}
\keyword{internal}
