% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/classes.R
\docType{class}
\name{dbR6Parent}
\alias{dbR6Parent}
\alias{dbR6Parent}
\title{dbR6_data class}
\format{\code{\link{R6Class}} object.}
\usage{
dbR6Parent
}
\value{
Object of \code{\link{R6Class}}
}
\description{
dbR6_data class
}
\section{Fields}{

\describe{
\item{\code{where}}{environment storing data & enclosing environment for metadata}
}}

\section{Methods}{

\describe{
\item{initialize}{initialize method}
\item{finalize}{finalize method}
\item{get_where}{get environment with the database connection}
\item{set_data}{set database connection}
}
}

\keyword{datasets}
