% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbR6Parent_set_metadata.R
\name{dbR6Parent_set_metadata}
\alias{dbR6Parent_set_metadata}
\title{dbR6Parent_set_metadata__}
\usage{
dbR6Parent_set_metadata(...)
}
\description{
dbR6Parent_set_metadata__
}
\keyword{internal}
