% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{generate_template}
\alias{generate_template}
\title{Generate template for dbR6 method}
\usage{
generate_template(external_fun, args = TRUE, what_class)
}
\description{
Generate template for dbR6 method
}
\keyword{internal}
