% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbR6_get_connection.R
\name{get_connection}
\alias{get_connection}
\alias{dbR6_get_connection}
\alias{get_connection,dbR6}
\title{add_table}
\usage{
dbR6_get_connection(...)
}
\description{
get sqlite connection
}
\examples{
{
data(mtcars2)
my_db <- dbR6$new()
my_db$add_table(mtcars, "mtcars_db")
}
}
