% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbR6_get_tables_number.R
\name{get_tables_number}
\alias{get_tables_number}
\alias{dbR6_get_tables_number}
\alias{get_tables_number,dbR6}
\title{get_tables_number}
\usage{
dbR6_get_tables_number()
}
\description{
get_tables_number
}
