% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbR6.R
\docType{class}
\name{dbR6}
\alias{dbR6}
\title{dbR6 class}
\format{\code{\link{R6Class}} object.}
\usage{
dbR6
}
\value{
Object of \code{\link{R6Class}}
}
\description{
dbR6 class
}
\section{Fields}{

\describe{
\item{\code{where}}{(inherited from R6_data class) Environment storing data & enclosing environment for metadata}

\item{\code{metadata}}{Environment storing object metadata}
}}

\section{Methods}{

\describe{
\item{deep_clone}{ipsus lorum}
\item{initialize}{ipsus lorum}
\item{finalize}{ipsus lorum}
\item{get_where}{(inherited from R6_data class) ipsus lorum}
\item{set_data}{(inherited from R6_data class) ipsus lorum}
\item{get_metadata}{ipsus lorum}
\item{set_metadata}{ipsus lorum}
\item{set_metadata_single}{ipsus lorum}
\item{list_tables}{ipsus lorum}
\item{set_metadata}{ipsus lorum}
\item{colnames}{ipsus lorum}
\item{nrow}{ipsus lorum}
\item{ncol}{ipsus lorum}
\item{print}{ipsus lorum}
\item{get_table}{ipsus lorum}
\item{send_query}{ipsus lorum}
\item{send_statement}{ipsus lorum}
\item{add_table}{ipsus lorum}
\item{remove_table}{ipsus lorum}
\item{add_empty_table}{ipsus lorum}
\item{save}{ipsus lorum}
\item{clone_db}{ipsus lorum}
\item{streamer}{ipsus lorum}
}
}

\keyword{datasets}
