% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbR6.R
\docType{class}
\name{dbR6}
\alias{dbR6}
\title{dbR6 class}
\format{\code{\link{R6Class}} object.}
\usage{
dbR6
}
\value{
Object of \code{\link{R6Class}}
}
\description{
dbR6 class
}
\section{Fields}{

\describe{
\item{\code{where}}{(inherited from R6_data class) Environment storing data & enclosing environment for metadata}

\item{\code{metadata}}{Environment storing object metadata}
}}

\section{Methods}{

\describe{
\item{deep_clone}{lorem ipsum}
\item{initialize}{lorem ipsum}
\item{finalize}{lorem ipsum}
\item{get_where}{(inherited from R6_data class) lorem ipsum}
\item{set_data}{(inherited from R6_data class) lorem ipsum}
\item{get_metadata}{lorem ipsum}
\item{set_metadata}{lorem ipsum}
\item{set_metadata_single}{lorem ipsum}
\item{list_tables}{lorem ipsum}
\item{set_metadata}{lorem ipsum}
\item{colnames}{lorem ipsum}
\item{nrow}{lorem ipsum}
\item{ncol}{lorem ipsum}
\item{dim}{lorem ipsum}
\item{print}{lorem ipsum}
\item{get_table}{lorem ipsum}
\item{send_query}{lorem ipsum}
\item{send_statement}{lorem ipsum}
\item{add_table}{lorem ipsum}
\item{remove_table}{lorem ipsum}
\item{add_empty_table}{lorem ipsum}
\item{save}{lorem ipsum}
\item{clone_db}{lorem ipsum}
\item{sort}{lorem ipsum}
\item{create_index}{lorem ipsum}
\item{drop_index}{lorem ipsum}
\item{get_indices}{lorem ipsum}
\item{filter}{lorem ipsum}
\item{exist_table}{lorem ipsum}
\item{streamer}{lorem ipsum}
}
}

\keyword{datasets}
